Title: 사각형 채우기 3, Time: 126ms, Memory: 10MB, Status: Passed - Codetree

+++ 250717/사각형 채우기 3/rectangle-fill-3.java
@@ -3,7 +3,7 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        int[] dp = new int[Math.max(n+1,3)];
+        long[] dp = new long[Math.max(n+1,3)];
         dp[0] = 1;
         dp[1] = 2;
